requirements.txt
Django>=4.2,<5.0
django-allauth>=0.57.0
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
